+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Possibility to send `application/json` payload with `HttpPostToJobNode`
 - Support for the EdDSA - ED25519 SSH key pair generation
 - Options `ConnectionRetryAttempts` and `ConnectionTimeout` for SSH client component are now configurable from `appsettings.json`
+- API HTTP Request logging with payload (redacted output on `Sensitive data`)
 - `Reason` attribute propagation from the HPC job (in the HEAppE Task)
 
 ### Fixed